LF-4220 update product select to be searchable
